+++ packages/ui/components/layout/index.tsx
@@ -1,4 +1,3 @@
 // codegen:start {preset: barrel, include: ./*.tsx, exclude: ./*.stories.tsx}
-export * from './BodyGrid'
 export * from './UserMenu'
 // codegen:end
