+++ packages/ui/layouts/index.tsx
@@ -1,3 +1,4 @@
 // codegen:start {preset: barrel, include: ./*.ts*, exclude: ./*.stories.*}
+export * from './BodyGrid'
 export * from './Organization'
 // codegen:end
